Add mention of always using --master
* Add to “Notes” at the bottom that the emperor should always be
  started with ``--master``
* Add ``--master`` to the invocation examples, so as not to suggest
  that invoking uwsgi without ``--master`` is generally appropriate
  for emperor instances

It is at least my understanding that the master is always required,
and reloads go horribly wrong if I don't enable it. But the docs are
vague about this, which is what I am trying to fix.

If there *are* cases where it is appropriate/desirable/necessary to
run the emperor without ``--master``, then please ignore this commit
but please also document the actual rules. :)