Improved explanation of the spooler file format
We had a bunch of failing spool files, and I needed to figure out what was in them. I am not *super* familiar with uWSGI. I didn't understand the documentation above which explained the file format. I ended up having to dig through the source code, which when you're totally unfamiliar with the codebase AND don't often (ever?) program in C, wasn't particularly easy.

Anyway, I thought it'd be nice to update the docs with a more full explanation of the file format.