max-requests-delta has not implemented yet
max-requests-delta seems not to be implemented on the latest release (2.0.18). I can not find max-requests-delta options on [uwsgi/core/uwsgi.c](https://github.com/unbit/uwsgi/blob/2.0.18/core/uwsgi.c)
If you configure max-requests-delta option on the latest release, it does not work without any warning.
It should be removed on the documents or add some notes to warn it does not work.